#21524 Moved {Stubbed,Effectful}ActorContext to testkit
